Decalred variables as double to avoid float-double conversion later in the loop.

+++ lib/Stokes_flow_Incomp.c
@@ -164,7 +164,7 @@ static float solve_Ahat_p_fhat_CG(struct All_variables *E,
     double global_vdot(), global_pdot();
 
     double time0, CPU_time0();
-    float dpressure, dvelocity;
+    double dpressure, dvelocity;
 
     void assemble_c_u();
     void assemble_div_u();
